+++ src/frontend/app/components/base/modal/styles.module.scss
@@ -2,3 +2,18 @@
     display: block;
     backdrop-filter: blur(7px);
 }
+
+.header {
+    background-color: #636668;
+    color: #ffffff;
+}
+
+.body {
+    background-color: #323539;
+    color: #ffffff;
+}
+
+.footer {
+    background-color: #323539;
+    color: #ffffff;
+}
