--- src/frontend/app/components/base/search_bar_input/index.js
@@ -1,1 +1,0 @@
-export {SearchBarInput} from './components';
