+++ src/frontend/app/components/main/section/search_panel/add_task_modal/styles.module.scss
@@ -1,3 +1,8 @@
 .form {
     cursor: default;
+    .title,
+    .status,
+    .description {
+        width: 90%;
+    }
 }
