+++ src/frontend/app/components/main/section/search_panel/search_bar_input/index.js
@@ -1,0 +1,1 @@
+export { SearchBarInput } from './components';
