+++ src/frontend/app/components/main/section/table/row/index.js
@@ -1,0 +1,1 @@
+export { Row } from "./components";
