+++ src/frontend/app/components/main/section/table/row/styles.module.scss
@@ -1,0 +1,3 @@
+.row:hover {
+    cursor: pointer;
+}
