+++ src/frontend/app/components/main/section/table/styles.module.scss
@@ -5,10 +5,6 @@
     border-color: #ffffff;
 }
 
-.row:hover {
-    cursor: pointer;
-}
-
 .heading-row {
     cursor: default;
 }
